+++ lock4j-core/src/main/java/com/baomidou/lock/DefaultLockFailureStrategy.java
@@ -19,16 +19,18 @@ package com.baomidou.lock;
 import com.baomidou.lock.exception.LockFailureException;
 import lombok.extern.slf4j.Slf4j;
 
+import java.lang.reflect.Method;
+
 /**
  * @author zengzhihong
  */
 @Slf4j
 public class DefaultLockFailureStrategy implements LockFailureStrategy {
 
+    protected static String DEFAULT_MESSAGE = "request failed,please retry it.";
+
     @Override
-    public void onLockFailure(String key, long acquireTimeout, int acquireCount) {
-        log.debug("thread:{} acquire lock fail,key:{} acquireTimeout:{}ms", Thread.currentThread().getName(), key,
-                acquireTimeout);
-        throw new LockFailureException("request failed,please retry it.");
+    public void onLockFailure(String key, Method method, Object[] arguments) {
+        throw new LockFailureException(DEFAULT_MESSAGE);
     }
 }