+++ lock4j-core/src/main/java/com/baomidou/lock/annotation/Lock4j.java
@@ -17,6 +17,7 @@
 package com.baomidou.lock.annotation;
 
 import com.baomidou.lock.executor.LockExecutor;
+import com.baomidou.lock.spring.boot.autoconfigure.Lock4jProperties;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -33,19 +34,28 @@ import java.lang.annotation.Target;
 public @interface Lock4j {
 
     /**
+     * 用于多个方法锁同一把锁 可以理解为锁资源名称 为空则会使用 包名+类名+方法名
+     *
+     * @return 名称
+     */
+    String name() default "";
+
+    /**
      * @return lock 执行器
      */
     Class<? extends LockExecutor> executor() default LockExecutor.class;
 
     /**
-     * @return KEY 默认包名+方法名
+     * support SPEL expresion 锁的key = name + keys
+     *
+     * @return KEY
      */
     String[] keys() default "";
 
     /**
      * @return 过期时间 单位：毫秒
      * <pre>
-     *     过期时间一定是要长于业务的执行时间. 未设置则为默认时间3秒
+     *     过期时间一定是要长于业务的执行时间. 未设置则为默认时间3秒 默认值：{@link Lock4jProperties#expire}
      * </pre>
      */
     long expire() default 0;
@@ -53,7 +63,7 @@ public @interface Lock4j {
     /**
      * @return 获取锁超时时间 单位：毫秒
      * <pre>
-     *     结合业务,建议该时间不宜设置过长,特别在并发高的情况下. 未设置则为默认时间3秒
+     *     结合业务,建议该时间不宜设置过长,特别在并发高的情况下. 未设置则为默认时间3秒 默认值：{@link Lock4jProperties#acquireTimeout}
      * </pre>
      */
     long acquireTimeout() default 0;