+++ lock4j-core/src/main/java/com/baomidou/lock/spring/boot/autoconfigure/LockAutoConfiguration.java
@@ -16,7 +16,11 @@
 
 package com.baomidou.lock.spring.boot.autoconfigure;
 
-import com.baomidou.lock.*;
+import com.baomidou.lock.DefaultLockFailureStrategy;
+import com.baomidou.lock.DefaultLockKeyBuilder;
+import com.baomidou.lock.LockFailureStrategy;
+import com.baomidou.lock.LockKeyBuilder;
+import com.baomidou.lock.LockTemplate;
 import com.baomidou.lock.aop.LockAnnotationAdvisor;
 import com.baomidou.lock.aop.LockInterceptor;
 import com.baomidou.lock.executor.LockExecutor;
@@ -44,10 +48,9 @@ public class LockAutoConfiguration {
     @SuppressWarnings("rawtypes")
     @Bean
     @ConditionalOnMissingBean
-    public LockTemplate lockTemplate(LockFailureStrategy lockFailureStrategy, List<LockExecutor> executors) {
+    public LockTemplate lockTemplate(List<LockExecutor> executors) {
         LockTemplate lockTemplate = new LockTemplate();
         lockTemplate.setProperties(properties);
-        lockTemplate.setLockFailureStrategy(lockFailureStrategy);
         lockTemplate.setExecutors(executors);
         return lockTemplate;
     }
@@ -66,8 +69,9 @@ public class LockAutoConfiguration {
 
     @Bean
     @ConditionalOnMissingBean
-    public LockInterceptor lockInterceptor(LockTemplate lockTemplate, LockKeyBuilder lockKeyBuilder) {
-        return new LockInterceptor(lockTemplate, lockKeyBuilder);
+    public LockInterceptor lockInterceptor(LockTemplate lockTemplate, LockKeyBuilder lockKeyBuilder,
+                                           LockFailureStrategy lockFailureStrategy) {
+        return new LockInterceptor(lockTemplate, lockKeyBuilder, lockFailureStrategy, properties);
     }
 
     @Bean