+++ lock4j-test/src/main/java/com/baomidou/lock/test/service/UserService.java
@@ -44,9 +44,15 @@ public class UserService {
 
     }
 
-    @Lock4j(keys = "#user.id")
+    @Lock4j(keys = "#user.id", acquireTimeout = 5000, expire = 5000)
     public User method1(User user) {
         System.out.println("执行spel方法1 , 当前线程:" + Thread.currentThread().getName() + " , counter：" + (counter++));
+        //模拟锁占用
+        try {
+            Thread.sleep(4000);
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
         return user;
     }